pyproject.toml
[project]
name = "ai-news-navigator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "groq>=1.1.1",
    "openai>=2.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
